mv set_plots_size_params from compare_pipelines to utils.plotting
