+++ Python_pipeline/compare_pipelines.py
@@ -13,11 +13,13 @@ import sys
 import pandas as pd
 from matplotlib import pyplot as plt
 from Join import wrangle_freqs_df
+
 STERNLAB_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(STERNLAB_PATH)
 from utils.logger import pipeline_logger
 from utils.pbs_jobs import create_pbs_cmd, submit
 from utils.runner_utils import submit_wait_and_log, FindFilesInDir
+from utils.plotting import set_plots_size_params
 
 
 def _create_python_output_folder(output_folder):
@@ -118,18 +120,6 @@ def _apply_invert_deletions(row, col):
         return row[col] * -1
     else:
         return row[col]
-
-
-def set_plots_size_params(size):
-    # Adapted from https://stackoverflow.com/questions/3899980/how-to-change-the-font-size-on-a-matplotlib-plot
-    bigger = size * 1.2
-    slightly_bigger = size * 1.1
-    plt.rc('font', size=size)                        # controls default text sizes
-    plt.rc('axes', titlesize=bigger)                 # fontsize of the axes title
-    plt.rc('axes', labelsize=slightly_bigger)        # fontsize of the x and y labels
-    plt.rc('xtick', labelsize=size)                  # fontsize of the tick labels
-    plt.rc('ytick', labelsize=size)                  # fontsize of the tick labels
-    plt.rc('legend', fontsize=size)                  # legend fontsize
 
 
 def plot_indels(data, output_folder):
