+++ utils/plotting.py
@@ -1,0 +1,13 @@
+from matplotlib import pyplot as plt
+
+
+def set_plots_size_params(size):
+    # Adapted from https://stackoverflow.com/questions/3899980/how-to-change-the-font-size-on-a-matplotlib-plot
+    bigger = size * 1.2
+    slightly_bigger = size * 1.1
+    plt.rc('font', size=size)                        # controls default text sizes
+    plt.rc('axes', titlesize=bigger)                 # fontsize of the axes title
+    plt.rc('axes', labelsize=slightly_bigger)        # fontsize of the x and y labels
+    plt.rc('xtick', labelsize=size)                  # fontsize of the tick labels
+    plt.rc('ytick', labelsize=size)                  # fontsize of the tick labels
+    plt.rc('legend', fontsize=size)                  # legend fontsize
